Implement optional xp command (#4)
* Implement Individual XP

* Implement additional XP features

- Added CONFIG_FLOAT_RATE_XP_PERSONAL_BONUS and CONFIG_UINT32_XP_PERSONAL_BONUS_REQ_ACCOUNT_LEVELconfig

- Added character export restriction method

* Misc

* Update Player.cpp

* Update CharacterHandler.cpp

* Update Player.cpp

* Update mangosd.conf.dist.in

+++ src/game/Entities/Unit.h
@@ -2398,7 +2398,7 @@ class Unit : public WorldObject
         void UpdateScalingAuras();
 
         uint32 GetDamageDoneByOthers() { return m_damageByOthers; }
-        uint32 GetModifierXpBasedOnDamageReceived(uint32 xp);
+        float GetModifierXpBasedOnDamageReceived(float xp);
         
         void OverrideMountDisplayId(uint32 newDisplayId);
